Refactor TLS check
